refactor: agent api intialiation
